Modify main page design
- add super tabs
- add timer, store page
