+++ src/pages/home/home.html
@@ -1,103 +1,14 @@
-<ion-header>
+<ion-header no-shadow no-border>
   <ion-navbar>
-    <ion-title>
-      Ionic Blank
-    </ion-title>
+    <button ion-button menuToggle color="yellow">
+      <ion-icon name="menu"></ion-icon>
+    </button>
+    <ion-title>{{pageTitle}}</ion-title>
   </ion-navbar>
 </ion-header>
-<ion-content>
-<super-tabs [rootNavCtrl]="navCtrl" [selectedTabIndex]="1"  height="100%" tabsColor="light" toolbarColor="dark" sliderColor="light">
-  <super-tab [tabRoot]="login" [title]="showTitles? 'First page' : ''" [icon]="showIcons? 'home': ''"></super-tab>
-  <super-tab [tabRoot]="page2" [title]="showTitles? 'Second page' : ''" [icon]="showIcons? 'pin' : ''"></super-tab>
-  <super-tab [tabRoot]="page3" [title]="showTitles? 'Third page' : ''" [icon]="showIcons? 'heart' : ''"></super-tab>
-</super-tabs>
+<ion-content no-bounce>
+  <super-tabs (tabSelect)="onTabSelect($event)">
+    <super-tab [root]="timer" title="Timer page"></super-tab>
+    <super-tab [root]="store" title="Store page"></super-tab>
+  </super-tabs>
 </ion-content>
-<!--
-<ion-content padding>
-
-  <button ion-button color="light" (click)="goLoginPage()">
-    login
-  </button>
-  <button ion-button color="light" (click)="goLogout()">
-    logout
-  </button>
-  <button ion-button color="light" (click)="goIncrease()">
-    timepicker
-  </button>
-  <button ion-button color="light" (click)="goStartFlow()">
-    start Flow
-  </button>
-  <button ion-button color="light" (click)="goTimerAction(btnState, '')">
-    {{btnState}}
-  </button>
-  <div>
-    current: {{current}}
-  </div>
-  <ion-item>
-    <ion-label>Date</ion-label>
-    <ion-datetime  displayFormat="HH:mm:ss" (ionChange)="onChangeTimeSet(timerString)" [(ngModel)]="timerString">
-    </ion-datetime>
-  </ion-item>
-
-  <ion-card *ngFor="let item of timerList ; let i = index" [attr.id]="'timerId_' + item.id">
-    <ion-item>
-      <p>{{item.title}} November 5, 1955</p>
-    </ion-item>
-
-
-    <div class="progress-wrapper" >
-      <div>
-        <i [style.color]=item.color class="icon-spin3 progress-refresh" (click)="goTimerReset(item)" ></i>
-
-        <ion-icon [style.color]=item.color class="progress-current" *ngIf="item.nextTimer==true" name="radio-button-on" (click)="goNextTimerStop()"></ion-icon>
-        <ion-icon [style.color]=item.color class="progress-current" *ngIf="item.nextTimer==false" name="radio-button-off" (click)="goNextTimerSet(item, i)"></ion-icon>
-
-        <ion-icon [style.color]=item.color class="progress-setting" name="settings" (click)="goTimerSetting(item)"></ion-icon>
-
-        <ion-icon *ngIf="item.notification.enable==true" [style.color]=item.color class="progress-notification" name="notifications" (click)="goNotificationToggle(item)"></ion-icon>
-        <ion-icon *ngIf="item.notification.enable==false" [style.color]=item.color class="progress-notification" name="notifications-off" (click)="goNotificationToggle(item)"></ion-icon>
-
-        <i *ngIf="item.status=='complete'" [style.color]=item.color class="icon-ok progress-status "></i>
-
-        <ion-spinner *ngIf="item.status=='running'" [style.stroke]=item.color class="progress-status" name="bubbles"></ion-spinner>
-
-        <div class="progress-text">
-          <ion-datetime [style.color]=item.color displayFormat="HH:mm:ss" (ionCancel)="onCancelTimeSet(item)" (ionChange)="onChangeTimeSet(item)" [(ngModel)]="item.timeSet">
-          </ion-datetime>
-        </div>
-        <div>
-          <i *ngIf="item.btnStatus=='start'" [style.color]=item.color class="icon-play progress-icon" (click)="goTimerAction(item)" ></i>
-          <i *ngIf="item.btnStatus=='pause'" [style.color]=item.color class="icon-pause progress-icon" (click)="goTimerAction(item)" ></i>
-          <i *ngIf="item.btnStatus=='resume'" [style.color]=item.color class="icon-play progress-icon" (click)="goTimerAction(item)"></i>
-          <i *ngIf="item.btnStatus=='end'" [style.color]=item.color class="icon-play progress-icon"></i>
-        </div>
-        <round-progress [current]="item.current" [max]="item.max" [radius]="radius" [rounded]="rounded" [color]=item.color >
-        </round-progress>
-      </div>
-    </div>
-
-    <ion-card-content>
-      <p>{{item.detail}}</p>
-    </ion-card-content>
-
-    <ion-row>
-      <ion-col>
-        <button ion-button icon-left clear small>
-          <ion-icon name="thumbs-up"></ion-icon>
-          <div>12 Likes</div>
-        </button>
-      </ion-col>
-      <ion-col>
-        <button ion-button icon-left clear small>
-          <ion-icon name="text"></ion-icon>
-          <div>4 Comments</div>
-        </button>
-      </ion-col>
-      <ion-col center text-center>
-        <ion-note>
-          11h ago
-        </ion-note>
-      </ion-col>
-    </ion-row>
-  </ion-card>
-</ion-content> -->
