+++ src/pages/store/store.html
@@ -1,10 +1,53 @@
-<ion-header>
-  <ion-navbar>
-    <ion-title>
-      Ionic Blank
-    </ion-title>
-  </ion-navbar>
-</ion-header>
-<ion-content>
+<ion-content padding>
+   <h1>
+     This is the first page
+   </h1>
+  <button ion-button (click)="pushPage()">Push page</button>
+  <button ion-button (click)="pushPage(true)">Push page in here</button>
 
+  <ion-slides slidesPerView="5">
+    <ion-slide>
+      <span class="color blue"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color red"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color yellow"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color blue"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color red"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color yellow"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color blue"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color red"></span>
+    </ion-slide>
+    <ion-slide>
+      <span class="color yellow"></span>
+    </ion-slide>
+  </ion-slides>
+
+  <p>
+    Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer et iaculis mauris. Nunc blandit viverra enim vehicula efficitur. Sed neque neque, tempus ac molestie non, blandit ut eros. Cras vulputate id nunc eget efficitur. Maecenas a dictum orci. Nullam sodales et est in euismod. Ut eget nibh consequat, molestie metus in, tristique lacus. Phasellus a elementum lectus. Morbi cursus mauris dapibus odio convallis convallis. Nulla et nisi nec augue consectetur rhoncus. Ut in tellus sit amet eros bibendum placerat ut a mi.
+  </p>
+  <p>
+    Quisque sagittis et nunc id dapibus. Curabitur placerat nec sem id hendrerit. Mauris tempus ante ac ante semper, non semper lacus maximus. Mauris in sem eros. Cras interdum nisl mollis magna lacinia, vitae vehicula enim ornare. Curabitur et lectus vel dolor venenatis pulvinar. Nullam malesuada vitae turpis at cursus.
+  </p>
+  <p>
+    Quisque turpis nisl, semper vel scelerisque sit amet, accumsan eu tellus. Maecenas non justo interdum, bibendum nisi eget, egestas enim. Aliquam suscipit ultrices magna, porttitor dapibus ipsum rhoncus at. Vivamus consectetur tortor a accumsan finibus. Aenean eleifend a tortor non porta. Sed iaculis eros sed lorem lacinia, ac imperdiet nisl molestie. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla facilisi. Nullam finibus condimentum eleifend. Pellentesque vel sapien et sem posuere ornare eu ac leo.
+  </p>
+  <p>
+    Praesent non ullamcorper turpis. Praesent eleifend est eget iaculis lobortis. Fusce elementum arcu lacus. Cras sed erat convallis, euismod augue at, euismod nisi. Morbi ac ligula rhoncus, venenatis sem et, posuere dui. Vivamus porta viverra massa, elementum vehicula diam euismod sed. Curabitur euismod justo magna, dapibus gravida est cursus sed. Ut nec blandit purus. Phasellus a auctor nibh.
+  </p>
+  <p>
+    Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia Curae; Ut faucibus mauris et sapien fermentum vehicula. Cras consectetur, nisi vel elementum viverra, turpis nibh tempor lacus, sit amet faucibus mi ex lobortis tellus. Aenean lobortis, augue et laoreet vehicula, arcu massa molestie mauris, vitae ullamcorper enim tortor vel ligula. Suspendisse quis facilisis leo, nec tempus tortor. Sed quis faucibus nisi. Cras molestie ornare aliquam. Integer faucibus commodo mauris a tempor. Aliquam rhoncus magna vulputate erat posuere sagittis. Vivamus tincidunt finibus ante quis condimentum. Vestibulum eget elit sed sapien suscipit aliquam ut nec ex. Orci varius natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus. Ut eget erat turpis. Nunc vitae maximus augue, et lacinia sapien.
+  </p>
 </ion-content>
